backend/requirements.txt
FastApi
Uvicorn
Pydantic
pytest
typing
mysql-connector-python